refactor(model-sync): move service files to nested models directory
- relocate modelSync to services/models/modelSync
- update all imports across background, tests, and services
- adjust relative paths in moved files
