+++ entrypoints/background/servicesInit.ts
@@ -2,7 +2,7 @@ import { autoRefreshService } from "~/services/accounts/autoRefreshService"
 import { autoCheckinScheduler } from "~/services/checkin/autoCheckin/scheduler"
 import { dailyBalanceHistoryScheduler } from "~/services/dailyBalanceHistory/scheduler"
 import { modelMetadataService } from "~/services/models/modelMetadata"
-import { modelSyncScheduler } from "~/services/modelSync"
+import { modelSyncScheduler } from "~/services/models/modelSync"
 import { redemptionAssistService } from "~/services/redemptionAssist"
 import { usageHistoryScheduler } from "~/services/usageHistory/scheduler"
 import { webdavAutoSyncService } from "~/services/webdav/webdavAutoSyncService"
