+++ tests/services/modelRedirect/ModelRedirectService.apply.test.ts
@@ -25,7 +25,7 @@ vi.mock("~/services/models/modelMetadata", () => ({
 const listChannelsMock = vi.fn()
 const updateChannelModelMappingMock = vi.fn()
 
-vi.mock("~/services/modelSync", () => {
+vi.mock("~/services/models/modelSync", () => {
   class ModelSyncServiceMock {
     listChannels = listChannelsMock
     updateChannelModelMapping = updateChannelModelMappingMock
